requirements.txt
streamlit==1.41.1
numpy==2.0.2
pandas==2.2.2
scipy==1.15.3
matplotlib==3.10.0
pillow==11.3.0
torch==2.8.0
fastai==2.8.4

# pip install -r requirements.txt

